cíclope en siete capas/requirements.txt
perplexity>=0.1.0
python-dotenv>=0.19.0
requests>=2.28.1
pytest>=7.0.0
pytest-mock>=3.10.0
